only show the member list when you have selected a room
